chore(toolkit-lib): cleanup unused  and private refactor exports
