+++ packages/aws-cdk/lib/api/refactor.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/no-relative-packages */
+export { mappingsByEnvironment, parseMappingGroups } from '../../../@aws-cdk/toolkit-lib/lib/actions/refactor/private/mapping-helpers';
